Zoom that follows the trackpad: fractional scroll deltas and an eased camera zoom target
pyglet reports trackpad scrolls as fractions of a line; truncating them to
int dropped light gestures entirely and turned a flick into several 1.1×
jumps. MouseEvent/InputEvent deltas are floats now, Camera gained
zoom_toward (an eased target anchored at a screen point, converging at a
frame-rate independent rate) and the map scales the target by 1.06 per
line, capped per event so momentum cannot overshoot. +/- ease as well.

+++ saga2d/backends/base.py
@@ -63,8 +63,8 @@ class MouseEvent:
     x: int
     y: int
     button: str | None = None
-    dx: int = 0
-    dy: int = 0
+    dx: float = 0.0  # drag: pointer movement; scroll: wheel lines, fractional on trackpads
+    dy: float = 0.0
 
 
 @dataclass(frozen=True)
